+++ pcs/cli/common/errors.py
@@ -1,7 +1,19 @@
 ERR_NODE_LIST_AND_ALL_MUTUALLY_EXCLUSIVE = (
     "Cannot specify both --all and a list of nodes."
 )
+SEE_MAN_CHANGES = "See 'man pcs' -> Changes in pcs-0.10."
+HINT_SYNTAX_CHANGE = (
+    "Syntax has changed from previous version. " + SEE_MAN_CHANGES
+)
 
+def msg_command_replaced(*new_commands):
+    new = "', '".join(new_commands)
+    return (
+        f"This command has been replaced with '{new}'. {SEE_MAN_CHANGES}"
+    )
+
+def raise_command_replaced(*new_commands):
+    raise CmdLineInputError(message=msg_command_replaced(*new_commands))
 
 class CmdLineInputError(Exception):
     """