+++ pcs/cli/routing/cluster.py
@@ -6,6 +6,7 @@ from pcs import (
     usage,
 )
 import pcs.cli.cluster.command as cluster_command
+from pcs.cli.common.errors import raise_command_replaced
 from pcs.cli.common.routing import create_router
 
 
@@ -69,6 +70,32 @@ cluster_cmd = create_router(
         "verify": cluster.cluster_verify,
         "report": cluster.cluster_report,
         "remove_nodes_from_cib": cluster.remove_nodes_from_cib,
+
+        # removed commands
+        # These print error messages which point users to the changes section in
+        # pcs manpage.
+        # To be removed in the next significant version.
+        "quorum": lambda lib, argv, modifiers: raise_command_replaced(
+            "pcs quorum"
+        ),
+        "remote-node": create_router(
+            {
+                "add": lambda lib, argv, modifiers: raise_command_replaced(
+                    "pcs cluster node add-guest",
+                ),
+                "remove": lambda lib, argv, modifiers: raise_command_replaced(
+                    "pcs cluster node delete-guest",
+                    "pcs cluster node remove-guest",
+                ),
+            },
+            ["cluster", "node"]
+        ),
+        "standby": lambda lib, argv, modifiers: raise_command_replaced(
+            "pcs node standby"
+        ),
+        "unstandby": lambda lib, argv, modifiers: raise_command_replaced(
+            "pcs node unstandby"
+        ),
     },
     ["cluster"]
 )