+++ pcs/prop.py
@@ -13,7 +13,13 @@ def set_property(lib, argv, modifiers):
       * -f - CIB file
     """
     del lib
-    modifiers.ensure_only_supported("--force", "-f")
+    modifiers.ensure_only_supported(
+        "--force", "-f",
+        # The hint is defined to print error messages which point users to the
+        # changes section in pcs manpage.
+        # To be removed in the next significant version.
+        hint_syntax_changed=modifiers.is_specified("--node")
+    )
     if not argv:
         raise CmdLineInputError()
 
@@ -68,7 +74,13 @@ def unset_property(lib, argv, modifiers):
       * -f - CIB file
     """
     del lib
-    modifiers.ensure_only_supported("--force", "-f")
+    modifiers.ensure_only_supported(
+        "--force", "-f",
+        # The hint is defined to print error messages which point users to the
+        # changes section in pcs manpage.
+        # To be removed in the next significant version.
+        hint_syntax_changed=modifiers.is_specified("--node")
+    )
     if not argv:
         raise CmdLineInputError()
 
@@ -85,7 +97,13 @@ def list_property(lib, argv, modifiers):
       * --defaults - list only default values of properties
     """
     del lib
-    modifiers.ensure_only_supported("--defaults", "--all", "-f")
+    modifiers.ensure_only_supported(
+        "--defaults", "--all", "-f",
+        # The hint is defined to print error messages which point users to the
+        # changes section in pcs manpage.
+        # To be removed in the next significant version.
+        hint_syntax_changed=modifiers.is_specified("--node")
+    )
     if len(argv) > 1:
         raise CmdLineInputError()
     print_all = len(argv) == 0